Lock file builder in seperate file
