+++ src/Microsoft.DotNet.ProjectModel/Graph/LockFile.cs
@@ -10,21 +10,6 @@ using NuGet.Versioning;
 
 namespace Microsoft.DotNet.ProjectModel.Graph
 {
-    public class LockFileBuilder
-    {
-        public string LockFilePath { get; set; }
-        public int Version { get; set; }
-        public IList<ProjectFileDependencyGroup> ProjectFileDependencyGroups { get; set; } = new List<ProjectFileDependencyGroup>();
-        public IList<LockFilePackageLibrary> PackageLibraries { get; set; } = new List<LockFilePackageLibrary>();
-        public IList<LockFileProjectLibrary> ProjectLibraries { get; set; } = new List<LockFileProjectLibrary>();
-        public IList<LockFileTarget> Targets { get; set; } = new List<LockFileTarget>();
-
-        public LockFile Build()
-        {
-            return new LockFile(LockFilePath, Version, ProjectFileDependencyGroups, PackageLibraries, ProjectLibraries, Targets);
-        }
-
-    }
     public class LockFile
     {
         public static readonly int CurrentVersion = 2;
